feat: check message when commit
